+++ src/finetune_lab/__init__.py
@@ -3,5 +3,5 @@
 Author: Asad Aslam
 """
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"
 __author__ = "Asad Aslam"